+++ test/test_provider.py
@@ -138,19 +138,19 @@ def test_remote_workspaces_table(httpx_mock: HTTPXMock) -> None:
     # List all available backends
     all_backends = provider.backends().by_workspace()
     assert set(all_backends) == {"default", "w1"}
-    assert {backend.resource_id for backend in all_backends["w1"]} == {"r1"}
-    assert {backend.resource_id for backend in all_backends["default"]} == {
+    assert {backend.resource_id.resource_id for backend in all_backends["w1"]} == {"r1"}
+    assert {backend.resource_id.resource_id for backend in all_backends["default"]} == {
         simulator.id for simulator in OFFLINE_SIMULATORS
     }
 
     # List only the devices
     only_devices = provider.backends(backend_type="device").by_workspace()
     assert set(only_devices) == {"w1"}
-    assert {backend.resource_id for backend in only_devices["w1"]} == {"r1"}
+    assert {backend.resource_id.resource_id for backend in only_devices["w1"]} == {"r1"}
 
     # List only the offline simulators
     only_offline_simulators = provider.backends(backend_type="offline_simulator").by_workspace()
     assert set(only_offline_simulators) == {"default"}
-    assert {backend.resource_id for backend in only_offline_simulators["default"]} == {
+    assert {backend.resource_id.resource_id for backend in only_offline_simulators["default"]} == {
         simulator.id for simulator in OFFLINE_SIMULATORS
     }