+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Check that the circuits submitted to the offline simulators can be converted to the AQT API (#68)
 * Update the user guide and improve the API reference consistency (#72, #75)
 * Add quickstart examples for the Qiskit.org homepage (#73)
+* Add persistence mechanism for `AQTJob` instances (#77)
+* Rename `OfflineSimulatorResource.noisy` to `OfflineSimulatorResource.with_noise_model` (#77)
 
 ## qiskit-aqt-provider v0.17.0